Extract admin links to server load function
