+++ src/routes/admin/+layout.svelte
@@ -1,24 +1,11 @@
 <script lang="ts">
 	import { page } from "$app/state";
 
-	const { children } = $props();
-
-	const pages = [
-		{
-			name: "Users",
-			href: "/admin/users",
-			icon: "person",
-		},
-		{
-			name: "Building Aliases",
-			href: "/admin/aliases",
-			icon: "menu_open",
-		},
-	];
+	const { data, children } = $props();
 </script>
 
 <nav class="tabbed">
-	{#each pages as pageLink}
+	{#each data.pages as pageLink}
 		<a class:active={page.url.pathname === pageLink.href} href={pageLink.href}>
 			<i>{pageLink.icon}</i>
 			<span>{pageLink.name}</span>
